requirements.txt
requests
pytest
pyyaml
allure-pytest